test/test-consumer/requirements.txt
-i https://pypi.org/simple
kafka-python==2.2.15